+++ ui/scripts/rules/__fixtures__/guncelleme/panel-sozlesmesiz/bad/kur-ornek.ps1
@@ -1,0 +1,4 @@
+$S = [hashtable]::Synchronized(@{ yuzde = 0 })
+$S.yuzde = 40
+$zaman = New-Object System.Windows.Forms.Timer
+$zaman.Interval = 250