+++ ui/scripts/rules/__fixtures__/guncelleme/panel-yok/bad/senkron.js
@@ -1,0 +1,7 @@
+const { execFile } = require('child_process');
+
+function esitle() {
+  return new Promise((coz) => execFile('git', ['pull'], (e) => coz(!e)));
+}
+
+module.exports = { esitle };